Create Module_1_Project.py
Project Problem Statement
To complete this module, you'll need to write a script that processes a bunch of images. It turns out that your company is in the process of updating its website, and a design contractor has been hired to create some new icon graphics for the site. However, the contractor has delivered the final designs and they’re in the wrong format, rotated 90° and too large. You’re unable to get in contact with the designers and your own deadline is approaching fast. You’ll need to use Python to get these images ready for launch!
So, how will you do this? You'll need to go through a folder full of images and operate with each of them. On each image, you'll use PIL methods like the ones we saw in the examples, and then write the new images in the right place.
If this sounds tricky, don't panic! You've already seen everything you need to do this, and now it's time to put it into practice.
As in the previous courses, the assessment will be done on a Virtual Machine running in the Cloud, thanks to the Qwiklabs infrastructure. You'll only have access to the VM for a limited amount of time, so we recommend that you write the script locally in your computer first, and only start the lab once your script is working correctly.
Good luck, you've got this!
Use the Python Imaging Library to do the following to a batch of images:
    Open an image
    Rotate an image
    Resize an image
    Save an image in a specific format in a separate directory 
You'll have 90 minutes to complete this lab.